pyproject.toml
[project]
name = "autotrade-data"
version = "0.1.0"
description = "IBKR market data collection pipeline — bars, news, breadth → Parquet + DuckDB"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ib_async>=1.0.0",
    "pyarrow>=15.0.0",
    "duckdb>=0.10.0",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer[all]>=0.9.0",
    "pandas-market-calendars>=4.3.0",
    "twikit>=2.0.0",
    "browser-cookie3>=0.20.1",
    "playwright>=1.58.0",
]

[project.scripts]
marketdata = "marketdata.cli:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["marketdata"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
